% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaccessR-package.R
\docType{data}
\name{census_data_cma_van}
\alias{census_data_cma_van}
\title{Boundary of the Vancouver Metropolitan Area.}
\format{
A simple features data frame with 1 row and 6 variables:
\describe{
\item{Type}{Type of statistical area: "CMA" for Census Metropolitan Area}
\item{GeoUID}{Unique identifier of the statistical area}
\item{name}{Name of the area}
\item{region}{Identifier of the region: "van" for Vancouver}
\item{study_region_name}{Name of the Census Metropolitan Area (CMA)}
\item{geometry}{Simple features geometry}
}
}
\source{
Population: Canada Census 2016. See: https://mountainmath.github.io/cancensus/index.html
}
\usage{
data(census_data_cma_van)
}
\description{
A simple features data frame with the boundary of the Metropolitan Area. The
boundary is unprojected (i.e., it is in latitude-longitude; CRS: WGS 84).
}
\examples{
 data(census_data_cma_van)
}
\keyword{datasets}
