% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaccessR-package.R
\docType{data}
\name{census_data_da_mtl}
\alias{census_data_da_mtl}
\title{Socio-economic and demographic data of the Montréal Metropolitan Area for Dissemination Areas.}
\format{
A simple features data frame with 6,469 row and 53 variables:
\describe{
\item{Type}{Type of statistical area: "DA" for Dissemination Area}
\item{GeoUID}{Unique identifier of the statistical area}
\item{region}{Identifier of the region: "mtl" for Montréal}
\item{Households}{Total number of households in dissemination area in 2016}
\item{CD_UID}{Unique identifier of the census CD}
\item{Dwellings}{Total number of dwellings in dissemination area in 2016}
\item{Population}{Total population in dissemination area in 2016}
\item{CSD_UID}{Unique identifier of the census CSD}
\item{CT_UID}{Unique identifier of the Census Tract}
\item{CMA_UID}{Unique identifier of the census Census Metropolitan Area}
\item{tot_age_popn}{Total age population}
\item{age_0_to_14}{Total population aged 0-14 years old in dissemination area in 2016}
\item{age_15_to_64}{Total population aged 15-64 years old in dissemination area in 2016}
\item{age_65_plus}{Total population aged 65 years and older in dissemination area in 2016}
\item{tot_cens_fam_priv_hh}{Total number of census families in private households in dissemination area in 2016}
\item{fem_parent}{define_fem_parent}
\item{LICO_AT}{define_lico_at}
\item{tot_LICO_status_2015}{define_tot_LICO}
\item{knwl_off_lang}{define_knwl_off_lang}
\item{neither_EN_FR}{define_neither_EN_FR}
\item{Aborin_ID_priv_HH}{define_Aborin_ID_priv_HH}
\item{Aborin_ID}{define_aborivin_ID}
\item{tot_priv_hh_hsing_suit}{define_tot_priv_hh_hsing_suit}
\item{not_suitable}{define_not_suitable}
\item{dwell_condtns}{define_dwell_condtns}
\item{maj_rep_needed}{define_maj_rep_needed}
\item{tot_shelt_cost}{define_tot_shelt_cost}
\item{shelt_gt30pc}{define_shelt_gt30pc}
\item{tot_tenant_HH}{define_tot_tenant_HH}
\item{tenant_HH_subsid}{define_tenant_HH_subsid}
\item{tot_immig_stat}{define_tot_immig_stat}
\item{immig_stat_last5}{define_immig_stat_last5}
\item{tot_adm_immig_popn_1980_to_2016}{define_tot_adm_immig_popn_1980_to_2016}
\item{refugees}{define_refugees}
\item{visible_min}{define_visible_min}
\item{tot_visible_min}{define_tot_visible_min}
\item{Black}{define_Black}
\item{tot_lbr_f_15_plus}{define_tot_lbr_f_15_plus}
\item{lbr_f_unemp}{define_lbr_f_unemp}
\item{tot_commuting_dur_lbr_f_15_plus}{define_tot_commuting_dur_lbr_f_15_plus}
\item{commuting_dur_60mins_plus}{define_commuting_dur_60mins_plus}
\item{from_mode_tran}{define_from_mode_tran}
\item{from_mode_walk}{define_from_mode_walk}
\item{ave_HH_size}{define_ave_HH_size}
\item{tot_degree_age_15_plus}{define_tot_degree_age_15_plus}
\item{bachelor_or_higher}{define_bachelor_or_higher}
\item{mhh_inc}{define_mhh_inc}
\item{from_mode_total}{define_from_mode_total}
\item{from_mode_car}{define_from_mode_car}
\item{from_mode_pass}{define_from_mode_pass}
\item{study_region_name}{Name of the Census Metropolitan Area (CMA)}
\item{Total_Emp}{define_Total_Emp}
\item{geometry}{Simple features geometry}
}
}
\source{
Population: Canada Census 2016. See: https://mountainmath.github.io/cancensus/index.html
}
\usage{
data(census_data_da_mtl)
}
\description{
A simple features data frame with the boundary of Dissemination Areas in the Metropolitan Area. The
boundary is unprojected (i.e., it is in latitude-longitude; CRS: WGS 84). The data frame contains
a selection of socio-economic and demographic information for the Dissemination Areas.
}
\examples{
 data(census_data_da_mtl)
}
\keyword{datasets}
