% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaccessR-package.R
\docType{data}
\name{population_statistics}
\alias{population_statistics}
\title{Population statistics for selected metropolitan areas in Canada.}
\format{
A data frame with 14 rows and 5 variables:
\describe{
\item{size_rank}{Rank of the metropolitan area according to population}
\item{year}{The date of the data}
\item{cma_name}{Name of the Census Metropolitan Area}
\item{province}{Name of the province that is home to the Census Metropolitan Area}
\item{population}{Population of the Census Metropolitan Area}
}
}
\source{
Statistics Canada.
}
\usage{
data(population_statistics)
}
\description{
A data frame with population for selected Census Metropolitan Areas in Canada
for 2021.
}
\examples{
 data(population_statistics)
 summary(population_statistics)
}
\keyword{datasets}
