% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaccessR-package.R
\docType{data}
\name{region_background_edm}
\alias{region_background_edm}
\title{Boundary for the background of Edmonton Metropolitan Area.}
\format{
A simple features data frame with 2 rows and 2 variables:
\describe{
\item{region}{Identifier of the region: "edm" for Edmonton}
\item{geometry}{Simple features geometry}
}
}
\usage{
data(region_background_edm)
}
\description{
A simple features data frame with the boundary of a polygon that can be used to
give a background to the Metropolitan Area. The boundary is unprojected (i.e.,
it is in latitude-longitude; CRS: WGS 84).
}
\examples{
 data(region_background_edm)
}
\keyword{datasets}
