% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaccessR-package.R
\docType{data}
\name{travel_matrix_emp_edm}
\alias{travel_matrix_emp_edm}
\title{Travel time matrix and employment statistics for Edmonton Metropolitan Area.}
\format{
A data frame with 2,057,486 rows and 9 variables:
\describe{
\item{from_id}{Unique identifier of the Dissemination Area of the zone of origin}
\item{to_id}{Unique identifier of the Dissemination Area of the zone of destination}
\item{travel_time_p50}{Median travel time in minutes by transit within a 15 minute window of the time of departure of 8:00 am}
\item{region}{Identifier of the region: "edm" for Edmonton}
\item{start_year}{Year for which travel time was calculated}
\item{start_datetime}{Date and time of departure for a trip that originates at the Dissemination Area of origin}
\item{study_region_name}{Name of the Census Metropolitan Area (CMA)}
\item{Population}{Total population in the Dissemination Area of origin (2016 Census)}
\item{Total_Emp}{Total employment in the Dissemination Area of destination (2016 Census)}
}
}
\source{
Population: Canada Census 2016. See: https://mountainmath.github.io/cancensus/index.html

Employment: Statistics Canada

Travel times: calculated based on General Transit Feed Standard and OpenStreetMaps data
}
\usage{
data(travel_matrix_emp_edm)
}
\description{
A data frame with origin-destination pairs, travel time by transit, and
population and employment by Census Dissemination Area. Census Dissemination
Areas (DAs) are the smallest publicly available geography provided by Statistics
Canada. Travel times are for a representative day in two different years:
2019 (pre-covid measures) and 2023 (post-covid measures) from the centroid of
the origin DA to the centroid of the destination DA. Times are calculated as
the median of all times within a 15 minutes window of the departure.
}
\examples{
 data(travel_matrix_emp_edm)
 summary(travel_matrix_emp_edm$travel_time_p50)
}
\keyword{datasets}
