% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaccessR-package.R
\docType{data}
\name{travel_matrix_grc_hal}
\alias{travel_matrix_grc_hal}
\title{Travel time to groceries for Halifax Metropolitan Area.}
\format{
A data frame with 50,747 rows and 8 variables:
\describe{
\item{from_id}{Unique identifier of the Dissemination Area of the zone of origin}
\item{to_id}{Unique identifier of the Dissemination Area of the zone of destination}
\item{travel_time_p50}{Median travel time in minutes by transit within a 15 minute window of the time of departure of 8:00 am}
\item{region}{Identifier of the region: "hal" for Halifax}
\item{start_datetime}{Date and time of departure for a trip that originates at the Dissemination Area of origin}
\item{start_year}{Year for which travel time was calculated}
\item{study_region_name}{Name of the Metropolita Area: Halifax}
\item{Population}{Total population in the Dissemination Area of origin (2016 Census)}
}
}
\source{
Population: Canada Census 2016. See: https://mountainmath.github.io/cancensus/index.html

Groceries: DMTI Points of Interest Data

Travel times: calculated based on General Transit Feed Standard and OpenStreetMaps data
}
\usage{
data(travel_matrix_grc_hal)
}
\description{
A data frame with origin-destination pairs, travel time by transit, and
population by Census Dissemination Area of origin. Dissemination Areas (DAs)
are the smallest publicly available geography provided by Statistics Canada.
Each destination is a single grocery opportunity, DEFINE THE SELECTION CRITERIA.
Travel times are for a representative day in two different years:
2019 (pre-covid measures) and 2023 (post-covid measures) from the centroid of the
origin DA to the address of the grocery. Times are calculated as the
median of all times within a 15 minutes window of the departure.
}
\examples{
 data(travel_matrix_grc_hal)
 summary(travel_matrix_grc_hal$travel_time_p50)
}
\keyword{datasets}
