% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaccessR-package.R
\docType{package}
\name{canaccessR-package}
\alias{canaccessR}
\alias{canaccessR-package}
\title{canaccessR: Data Package For Transit Accessibility Analysis in Selected Canadian Metropolitan Areas}
\description{
An open data product with travel times tables by transit on a representative weekday of 2019 (pre-pandemic measures) and 2023 (post-pandemic measures) for the analysis of employment accessibility, and travel time tables by transit on a representative weekend day on the same years for the analysis of grocery accessibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/canaccessR}
  \item \url{https://paezha.github.io/canaccessR/}
  \item Report bugs at \url{https://github.com/paezha/canaccessR/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Robert N. Arku \email{soukhoa@mcmaster.ca} (\href{https://orcid.org/0000-0002-2018-886X}{ORCID})
  \item Christopher D. Higgins \email{soukhoa@mcmaster.ca} (\href{https://orcid.org/0000-0002-3551-7750}{ORCID})
  \item João Pedro Parga \email{joao.parga@mail.utoronto.ca} (\href{https://orcid.org/0000-0002-4105-5927}{ORCID})
  \item Anastasia Soukhov \email{soukhoa@mcmaster.ca} (\href{https://orcid.org/0000-0003-4371-4831}{ORCID})
}

}
\keyword{internal}
